Refactor API handlers and add general topic creation endpoint with handler updated manager - ready for review
